Yet another attempt to fix the automated test cases
