+++ src/ShaderGen.Tests/TestAssets/PartialClassShader/PartialVertex1.cs
@@ -5,7 +5,7 @@ namespace TestShaders
 {
     public partial class PartialVertex
     {
-        struct FragmentInput
+        public struct FragmentInput
         {
             [VertexSemantic(SemanticType.SystemPosition)] public Vector4 Position;
             [VertexSemantic(SemanticType.Color)] public Vector4 Color;
@@ -14,7 +14,7 @@ namespace TestShaders
         public SamplerResource Sampler;
 
         [VertexShader]
-        FragmentInput VertexShaderFunc(VertexInput input)
+        public FragmentInput VertexShaderFunc(VertexInput input)
         {
             FragmentInput output;
             output.Position = new Vector4(input.Position, 1);
